requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
idna==3.4
Pillow==10.0.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
